refactor: move CreateTransactionDSL to transaction.go, delete empty file :hammer:
