+++ components/ledger/internal/adapters/http/in/transaction.go
@@ -5,11 +5,20 @@
 package in
 
 import (
-	libCommons "github.com/LerianStudio/lib-commons/v4/commons"
+	"fmt"
+	"reflect"
 
+	libCommons "github.com/LerianStudio/lib-commons/v4/commons"
+	libLog "github.com/LerianStudio/lib-commons/v4/commons/log"
+	libOpentelemetry "github.com/LerianStudio/lib-commons/v4/commons/opentelemetry"
+
+	"github.com/LerianStudio/midaz/v3/components/ledger/internal/adapters/postgres/transaction"
 	"github.com/LerianStudio/midaz/v3/components/ledger/internal/services/command"
 	"github.com/LerianStudio/midaz/v3/components/ledger/internal/services/query"
+	"github.com/LerianStudio/midaz/v3/pkg"
 	"github.com/LerianStudio/midaz/v3/pkg/constant"
+	goldTransaction "github.com/LerianStudio/midaz/v3/pkg/gold/transaction"
+	"github.com/LerianStudio/midaz/v3/pkg/net/http"
 	pkgTransaction "github.com/LerianStudio/midaz/v3/pkg/transaction"
 
 	"github.com/gofiber/fiber/v2"
@@ -170,4 +179,86 @@ func (handler *TransactionHandler) CreateTransactionOutflow(p any, c *fiber.Ctx)
 	recordSafePayloadAttributes(span, transactionInput)
 
 	return handler.createTransaction(c, *transactionInput, transactionInput.InitialStatus())
+}
+
+// CreateTransactionDSL method that create transaction using DSL
+//
+//	@Summary		Create a Transaction using DSL
+//	@Description	Create a Transaction with the input DSL file
+//	@Tags			Transactions
+//	@Accept			mpfd
+//	@Produce		json
+//	@Param			Authorization	header		string	true	"Authorization Bearer Token"
+//	@Param			X-Request-Id	header		string	false	"Request ID"
+//	@Param			organization_id	path		string	true	"Organization ID"
+//	@Param			ledger_id		path		string	true	"Ledger ID"
+//	@Param			transaction		formData	file	true	"Transaction DSL file"
+//	@Success		200				{object}	Transaction
+//	@Failure		400				{object}	mmodel.Error	"Invalid DSL file format or validation errors"
+//	@Failure		401				{object}	mmodel.Error	"Unauthorized access"
+//	@Failure		403				{object}	mmodel.Error	"Forbidden access"
+//	@Failure		422				{object}	mmodel.Error	"Unprocessable Entity, validation errors"
+//	@Failure		500				{object}	mmodel.Error	"Internal server error"
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/dsl [post]
+func (handler *TransactionHandler) CreateTransactionDSL(c *fiber.Ctx) error {
+	ctx := c.UserContext()
+
+	logger, tracer, _, _ := libCommons.NewTrackingFromContext(ctx)
+
+	ctx, span := tracer.Start(ctx, "handler.create_transaction_dsl")
+	defer span.End()
+
+	c.SetUserContext(ctx)
+
+	c.Set("Deprecation", "true")
+	c.Set("Sunset", "Sat, 01 Aug 2026 00:00:00 GMT")
+	c.Set("Link", "</v1/organizations/"+c.Params("organization_id")+
+		"/ledgers/"+c.Params("ledger_id")+
+		"/transactions/json>; rel=\"successor-version\"")
+
+	logger.Log(ctx, libLog.LevelWarn, fmt.Sprintf("DEPRECATED ENDPOINT: POST /transactions/dsl called. "+
+		"request_id=%s. Use POST /transactions/json instead. "+
+		"Sunset date: 2026-08-01", c.Get("X-Request-Id")))
+
+	_, err := http.ValidateParameters(c.Queries())
+	if err != nil {
+		libOpentelemetry.HandleSpanBusinessErrorEvent(span, "Failed to validate query parameters", err)
+
+		logger.Log(ctx, libLog.LevelError, fmt.Sprintf("Failed to validate query parameters: %s", err.Error()))
+
+		return http.WithError(c, err)
+	}
+
+	dsl, err := http.GetFileFromHeader(c)
+	if err != nil {
+		libOpentelemetry.HandleSpanBusinessErrorEvent(span, "Failed to get file from Header", err)
+
+		logger.Log(ctx, libLog.LevelError, fmt.Sprintf("Failed to get file from Header: %s", err.Error()))
+
+		return http.WithError(c, err)
+	}
+
+	errListener := goldTransaction.Validate(dsl)
+	if errListener != nil && len(errListener.Errors) > 0 {
+		err := pkg.ValidateBusinessError(constant.ErrInvalidDSLFileFormat, reflect.TypeOf(transaction.Transaction{}).Name(), errListener.Errors)
+
+		libOpentelemetry.HandleSpanBusinessErrorEvent(span, "Failed to validate script in DSL", err)
+
+		return http.WithError(c, err)
+	}
+
+	parsed := goldTransaction.Parse(dsl)
+
+	transactionInput, ok := parsed.(pkgTransaction.Transaction)
+	if !ok {
+		err := pkg.ValidateBusinessError(constant.ErrInvalidDSLFileFormat, reflect.TypeOf(transaction.Transaction{}).Name())
+
+		libOpentelemetry.HandleSpanBusinessErrorEvent(span, "Failed to parse script in DSL", err)
+
+		return http.WithError(c, err)
+	}
+
+	recordSafePayloadAttributes(span, transactionInput.Send)
+
+	return handler.createTransaction(c, transactionInput, transactionInput.InitialStatus())
 }
